change in directories and files (#30)
* Enhancement in GUI

Now the user doesn't need to clear the data that are present initially in the entry box. First time, the entry box clears itself.

* Update screenrec.py

* Update screenrec.py

* Delete screenrec.py

* added sim_btw_pics.py

Code file that compares the two images and calculates the difference between the two arrays of pixel values (two images)

* Update sim_btw_pics.py

* added images for sample comparison

* Update sim_btw_pics.py

* Delete sample1.png

* Delete sample2.png

* Delete sim_btw_pics.py

* change in directories

moved the .py file to utils folder and deleted the sample picture files.